resurrect secondary detectors
This change brings back the 'secondary detector' functionality from
MiniCLEAN RAT (and others), allowing users to create generic chambers
that register hits. This is useful for example for a veto or
scintillator counter, which is not fully simulated -- MC hit information
is stored in the DS in mc.secondaryhits for later processing.

The secondary detector hits are stored in a list (mc.secondaryhits) as
DS::MCSecondaryHit objects. These have the name of the volume and (since
that might not be unique) a unique numerical ID which can be used to,
e.g., pick a particular counter out of an array. The object also stores
the hit position, global time, total energy deposited, and the PDG code
of the interacting particle.

Editor's note: the classes GenericChamber(Hit) were renamed from
BWVetGenericChamber(Hit) and cleaned up a bit; these were in the code
but not used anywhere. MCSecondaryHit and the code to populate it were
ported from MiniCLEAN RAT.